pyproject.toml
[tool.poetry]
name = "pomodoro-backend"
version = "0.1.0"
description = "Pomodoro Timer"
authors = ["naresh-khatri <naresh.khatri2345@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2"
djangorestframework = "^3.12.4"
django-cors-headers = "^3.7.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
